wip: clean up old kdf/group primitives
