+++ manta-crypto/src/lib.rs
@@ -28,11 +28,9 @@ pub mod accumulator;
 pub mod algebra;
 pub mod commitment;
 pub mod constraint;
-// TODO: pub mod ecc;
 pub mod eclair;
 pub mod encryption;
 pub mod hash;
-// TODO: pub mod old_key;
 pub mod key;
 pub mod password;
 pub mod permutation;
